+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## v3.5.0 - 2026-05-26
+
 - fix: A11y issues on table of donations
 - feat: Allow use of clean theme